Handle required missing fields failures
At the moment, whenever a required field is missing the program crashes
as I wrongfully thought that force-unwraping an Optional would throw an
exception that would be propagated upward, which turns out not being the
case. Therefore, here I introduce 'StorkDecodeError' to throw a
type-safe (in the exceptions world) error that can be handled later. On
top of these changes, I here add new unit tests to verify this property
of Stork and pruned cases where 'throw' is defined for functions that
should not need to throw.
See https://github.com/NunoAlexandre/stork