Fix decode of required arrays on typemismatch
When a field x is expected to contain an Array but it contains some
other possible json value, Stork now returns an empty string. Instead,
it should fail. Note the unit tests adjusted and added in this commit to
cover this case and the case where the expected type of the (required)
array elements is not what's expected. For example, you expect [String]
but it actually is [Int]. Stork, in this last case, still returns an
empty list back. I will think about this behaviour. This happens because
it compactMap's decodeValue on the array, taking our the failing cases,
which are all the elements in type of such typemismatch.