Fix single-digit integers considered as Bool
Note that this is unit-tested where the starting point is Swift
Dictionaries. When the json actually comes from a file and therefore
goes through coercion, '1' and '0' int values fall into the Bool type
casting match, which is wrong. Here I first check if a value is an int
and only if not, I check if it's a Bool. For a better and more complete
fix I should drop covering .int and .double json values to have a single
.number case, which should probably be an NSNumber.