Drop JsonValue int and double in favour of number
When parsing real json, this data has potentially been coerced to 'Data'
and then / or as a JSON, i.e., [String: Any], which causes Boolean's and
Int's to get confused through the NSNumber type. See the following
stackOverflow thread to learn more about this problem:
http://tinyurl.com/y2u9zmxw.

These changes ensure that such confusion does not take place and adds
integration tests to verify that property.